requirements.txt
Flask==3.0.2
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Limiter==3.5.1
requests==2.31.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Werkzeug==3.0.1
Flask-Mail==0.10.0
cryptography==43.0.1
python-dateutil
watchdog
pandas
statsmodels
scikit-learn